Module google_ml1::api

Structs

CloudMachineLearningEngine
Central instance to access all CloudMachineLearningEngine related resource activities

GoogleApi__HttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page. This message can be used both in streaming and non-streaming API methods in the request as well as the response. It can be used as a top-level request field, which is convenient if one wants to extract parameters from either the URL or HTTP template into the request fields and also want access to the raw HTTP body. Example: message GetResourceRequest { // A unique request id. string request_id = 1; // The raw HTTP body is bound to this field. google.api.HttpBody http_body = 2; } service ResourceService { rpc GetResource(GetResourceRequest) returns (google.api.HttpBody); rpc UpdateResource(google.api.HttpBody) returns (google.protobuf.Empty); } Example with streaming methods: service CaldavService { rpc GetCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); rpc UpdateCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); } Use of this type only changes how the request and response bodies are handled, all other features will continue to work unchanged.

GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig
There is no detailed description.

GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig
The median automated stopping rule stops a pending trial if the trial's best objective_value is strictly below the median 'performance' of all completed trials reported up to the trial's last measurement. Currently, 'performance' refers to the running average of the objective values reported by the trial in each measurement.

GoogleCloudMlV1_HyperparameterOutput_HyperparameterMetric
An observed value of a metric.

GoogleCloudMlV1_Measurement_Metric
A message representing a metric in the measurement.

GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec
There is no detailed description.

GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec
There is no detailed description.

GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec
There is no detailed description.

GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec
There is no detailed description.

GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec
Represents the spec to match categorical values from parent parameter.

GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec
Represents the spec to match discrete values from parent parameter.

GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec
Represents the spec to match integer values from parent parameter.

GoogleCloudMlV1_StudyConfig_MetricSpec
Represents a metric to optimize.

GoogleCloudMlV1_StudyConfig_ParameterSpec
Represents a single parameter to optimize.

GoogleCloudMlV1_Trial_Parameter
A message representing a parameter to be tuned. Contains the name of the parameter and the suggested value to use for this trial.

GoogleCloudMlV1__AcceleratorConfig
Represents a hardware accelerator request config. Note that the AcceleratorConfig can be used in both Jobs and Versions. Learn more about accelerators for training and accelerators for online prediction.

GoogleCloudMlV1__AddTrialMeasurementRequest
The request message for the AddTrialMeasurement service method.

GoogleCloudMlV1__AutoScaling
Options for automatically scaling a model.

GoogleCloudMlV1__AutomatedStoppingConfig
Configuration for Automated Early Stopping of Trials. If no implementation_config is set, automated early stopping will not be run.

GoogleCloudMlV1__BuiltInAlgorithmOutput
Represents output related to a built-in algorithm Job.

GoogleCloudMlV1__CancelJobRequest
Request message for the CancelJob method.

GoogleCloudMlV1__Capability
There is no detailed description.

GoogleCloudMlV1__CheckTrialEarlyStoppingStateRequest
The request message for the CheckTrialEarlyStoppingState service method.

GoogleCloudMlV1__CompleteTrialRequest
The request message for the CompleteTrial service method.

GoogleCloudMlV1__Config
There is no detailed description.

GoogleCloudMlV1__ContainerPort
Represents a network port in a single container. This message is a subset of the Kubernetes ContainerPort v1 core specification.

GoogleCloudMlV1__ContainerSpec
Specification of a custom container for serving predictions. This message is a subset of the Kubernetes Container v1 core specification.

GoogleCloudMlV1__DiskConfig
Represents the config of disk options.

GoogleCloudMlV1__EncryptionConfig
Represents a custom encryption key configuration that can be applied to a resource.

GoogleCloudMlV1__EnvVar
Represents an environment variable to be made available in a container. This message is a subset of the Kubernetes EnvVar v1 core specification.

GoogleCloudMlV1__ExplainRequest
Request for explanations to be issued against a trained model.

GoogleCloudMlV1__ExplanationConfig
Message holding configuration options for explaining model predictions. There are three feature attribution methods supported for TensorFlow models: integrated gradients, sampled Shapley, and XRAI. Learn more about feature attributions.

GoogleCloudMlV1__GetConfigResponse
Returns service account information associated with a project.

GoogleCloudMlV1__HyperparameterOutput
Represents the result of a single hyperparameter tuning trial from a training job. The TrainingOutput object that is returned on successful completion of a training job with hyperparameter tuning includes a list of HyperparameterOutput objects, one for each successful trial.

GoogleCloudMlV1__HyperparameterSpec
Represents a set of hyperparameters to optimize.

GoogleCloudMlV1__IntegratedGradientsAttribution
Attributes credit by computing the Aumann-Shapley value taking advantage of the model's fully differentiable structure. Refer to this paper for more details: https://arxiv.org/abs/1703.01365

GoogleCloudMlV1__Job
Represents a training or prediction job.

GoogleCloudMlV1__ListJobsResponse
Response message for the ListJobs method.

GoogleCloudMlV1__ListLocationsResponse
There is no detailed description.

GoogleCloudMlV1__ListModelsResponse
Response message for the ListModels method.

GoogleCloudMlV1__ListOptimalTrialsRequest
The request message for the ListTrials service method.

GoogleCloudMlV1__ListOptimalTrialsResponse
The response message for the ListOptimalTrials method.

GoogleCloudMlV1__ListStudiesResponse
There is no detailed description.

GoogleCloudMlV1__ListTrialsResponse
The response message for the ListTrials method.

GoogleCloudMlV1__ListVersionsResponse
Response message for the ListVersions method.

GoogleCloudMlV1__Location
There is no detailed description.

GoogleCloudMlV1__ManualScaling
Options for manually scaling a model.

GoogleCloudMlV1__Measurement
A message representing a measurement.

GoogleCloudMlV1__MetricSpec
MetricSpec contains the specifications to use to calculate the desired nodes count when autoscaling is enabled.

GoogleCloudMlV1__Model
Represents a machine learning solution. A model can have multiple versions, each of which is a deployed, trained model ready to receive prediction requests. The model itself is just a container.

GoogleCloudMlV1__ParameterSpec
Represents a single hyperparameter to optimize.

GoogleCloudMlV1__PredictRequest
Request for predictions to be issued against a trained model.

GoogleCloudMlV1__PredictionInput
Represents input parameters for a prediction job.

GoogleCloudMlV1__PredictionOutput
Represents results of a prediction job.

GoogleCloudMlV1__ReplicaConfig
Represents the configuration for a replica in a cluster.

GoogleCloudMlV1__RequestLoggingConfig
Configuration for logging request-response pairs to a BigQuery table. Online prediction requests to a model version and the responses to these requests are converted to raw strings and saved to the specified BigQuery table. Logging is constrained by BigQuery quotas and limits. If your project exceeds BigQuery quotas or limits, AI Platform Prediction does not log request-response pairs, but it continues to serve predictions. If you are using continuous evaluation, you do not need to specify this configuration manually. Setting up continuous evaluation automatically enables logging of request-response pairs.

GoogleCloudMlV1__RouteMap
Specifies HTTP paths served by a custom container. AI Platform Prediction sends requests to these paths on the container; the custom container must run an HTTP server that responds to these requests with appropriate responses. Read Custom container requirements for details on how to create your container image to meet these requirements.

GoogleCloudMlV1__SampledShapleyAttribution
An attribution method that approximates Shapley values for features that contribute to the label being predicted. A sampling strategy is used to approximate the value rather than considering all subsets of features.

GoogleCloudMlV1__Scheduling
All parameters related to scheduling of training jobs.

GoogleCloudMlV1__SetDefaultVersionRequest
Request message for the SetDefaultVersion request.

GoogleCloudMlV1__StopTrialRequest
There is no detailed description.

GoogleCloudMlV1__Study
A message representing a Study.

GoogleCloudMlV1__StudyConfig
Represents configuration of a study.

GoogleCloudMlV1__SuggestTrialsRequest
The request message for the SuggestTrial service method.

GoogleCloudMlV1__TrainingInput
Represents input parameters for a training job. When using the gcloud command to submit your training job, you can specify the input parameters as command-line arguments and/or in a YAML configuration file referenced from the --config command-line argument. For details, see the guide to submitting a training job.

GoogleCloudMlV1__TrainingOutput
Represents results of a training job. Output only.

GoogleCloudMlV1__Trial
A message representing a trial.

GoogleCloudMlV1__Version
Represents a version of the model. Each version is a trained model deployed in the cloud, ready to handle prediction requests. A model can have multiple versions. You can get information about all of the versions of a given model by calling projects.models.versions.list.

GoogleCloudMlV1__XraiAttribution
Attributes credit by computing the XRAI taking advantage of the model's fully differentiable structure. Refer to this paper for more details: https://arxiv.org/abs/1906.02825 Currently only implemented for models with natural image inputs.

GoogleIamV1__AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both allServices and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { "audit_configs": [ { "service": "allServices", "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" }, { "log_type": "ADMIN_READ" } ] }, { "service": "sampleservice.googleapis.com", "audit_log_configs": [ { "log_type": "DATA_READ" }, { "log_type": "DATA_WRITE", "exempted_members": [ "user:aliya@example.com" ] } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts jose@example.com from DATA_READ logging, and aliya@example.com from DATA_WRITE logging.

GoogleIamV1__AuditLogConfig
Provides the configuration for logging a type of permissions. Example: { "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" } ] } This enables 'DATA_READ' and 'DATA_WRITE' logging, while exempting jose@example.com from DATA_READ logging.

GoogleIamV1__Binding
Associates members with a role.

GoogleIamV1__Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

GoogleIamV1__SetIamPolicyRequest
Request message for SetIamPolicy method.

GoogleIamV1__TestIamPermissionsRequest
Request message for TestIamPermissions method.

GoogleIamV1__TestIamPermissionsResponse
Response message for TestIamPermissions method.

GoogleLongrunning__ListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunning__Operation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobuf__Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpc__Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

GoogleType__Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

ProjectExplainCall
Performs explanation on the data in the request. {% dynamic include "/ai-platform/includes/___explain-request" %}

ProjectGetConfigCall
Get the service account information associated with your project. You need this information in order to grant the service account permissions for the Google Cloud Storage location where you put your model training code for training the model with Google Cloud Machine Learning.

ProjectJobCancelCall
Cancels a running job.

ProjectJobCreateCall
Creates a training or a batch prediction job.

ProjectJobGetCall
Describes a job.

ProjectJobGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectJobListCall
Lists the jobs in the project. If there are no jobs that match the request parameters, the list request returns an empty response body: {}.

ProjectJobPatchCall
Updates a specific job resource. Currently the only supported fields to update are labels.

ProjectJobSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

ProjectJobTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectLocationGetCall
Get the complete list of CMLE capabilities in a location, along with their location-specific properties.

ProjectLocationListCall
List all locations that provides at least one type of CMLE capability.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationStudyCreateCall
Creates a study.

ProjectLocationStudyDeleteCall
Deletes a study.

ProjectLocationStudyGetCall
Gets a study.

ProjectLocationStudyListCall
Lists all the studies in a region for an associated project.

ProjectLocationStudyTrialAddMeasurementCall
Adds a measurement of the objective metrics to a trial. This measurement is assumed to have been taken before the trial is complete.

ProjectLocationStudyTrialCheckEarlyStoppingStateCall
Checks whether a trial should stop or not. Returns a long-running operation. When the operation is successful, it will contain a CheckTrialEarlyStoppingStateResponse.

ProjectLocationStudyTrialCompleteCall
Marks a trial as complete.

ProjectLocationStudyTrialCreateCall
Adds a user provided trial to a study.

ProjectLocationStudyTrialDeleteCall
Deletes a trial.

ProjectLocationStudyTrialGetCall
Gets a trial.

ProjectLocationStudyTrialListCall
Lists the trials associated with a study.

ProjectLocationStudyTrialListOptimalTrialCall
Lists the pareto-optimal trials for multi-objective study or the optimal trials for single-objective study. The definition of pareto-optimal can be checked in wiki page. https://en.wikipedia.org/wiki/Pareto_efficiency

ProjectLocationStudyTrialStopCall
Stops a trial.

ProjectLocationStudyTrialSuggestCall
Adds one or more trials to a study, with parameter values suggested by AI Platform Vizier. Returns a long-running operation associated with the generation of trial suggestions. When this long-running operation succeeds, it will contain a SuggestTrialsResponse.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudMachineLearningEngine hub.

ProjectModelCreateCall
Creates a model which will later contain one or more versions. You must add at least one version before you can request predictions from the model. Add versions by calling projects.models.versions.create.

ProjectModelDeleteCall
Deletes a model. You can only delete a model if there are no versions in it. You can delete versions by calling projects.models.versions.delete.

ProjectModelGetCall
Gets information about a model, including its name, the description (if set), and the default version (if at least one version of the model has been deployed).

ProjectModelGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectModelListCall
Lists the models in a project. Each project can contain multiple models, and each model can have multiple versions. If there are no models that match the request parameters, the list request returns an empty response body: {}.

ProjectModelPatchCall
Updates a specific model resource. Currently the only supported fields to update are description and default_version.name.

ProjectModelSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

ProjectModelTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectModelVersionCreateCall
Creates a new version of a model from a trained TensorFlow model. If the version created in the cloud by this call is the first deployed version of the specified model, it will be made the default version of the model. When you add a version to a model that already has one or more versions, the default version does not automatically change. If you want a new version to be the default, you must call projects.models.versions.setDefault.

ProjectModelVersionDeleteCall
Deletes a model version. Each model can have multiple versions deployed and in use at any given time. Use this method to remove a single version. Note: You cannot delete the version that is set as the default version of the model unless it is the only remaining version.

ProjectModelVersionGetCall
Gets information about a model version. Models can have multiple versions. You can call projects.models.versions.list to get the same information that this method returns for all of the versions of a model.

ProjectModelVersionListCall
Gets basic information about all the versions of a model. If you expect that a model has many versions, or if you need to handle only a limited number of results at a time, you can request that the list be retrieved in batches (called pages). If there are no versions that match the request parameters, the list request returns an empty response body: {}.

ProjectModelVersionPatchCall
Updates the specified Version resource. Currently the only update-able fields are description, requestLoggingConfig, autoScaling.minNodes, and manualScaling.nodes.

ProjectModelVersionSetDefaultCall
Designates a version to be the default for the model. The default version is used for prediction requests made against the model that don't specify a version. The first version to be created for a model is automatically set as the default. You must make any subsequent changes to the default version setting manually using this method.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectPredictCall
Performs online prediction on the data in the request. {% dynamic include "/ai-platform/includes/___predict-request" %}

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.