Crate google_ml1

This documentation was generated from Cloud Machine Learning Engine crate version 1.0.9+20190621, where 20190621 is the exact revision of the ml:v1 schema built by the mako code generator v1.0.9.

Everything else about the Cloud Machine Learning Engine v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

get config, jobs cancel, jobs create, jobs get, jobs get iam policy, jobs list, jobs patch, jobs set iam policy, jobs test iam permissions, locations get, locations list, models create, models delete, models get, models get iam policy, models list, models patch, models set iam policy, models test iam permissions, models versions create, models versions delete, models versions get, models versions list, models versions patch, models versions set default, operations cancel, operations get, operations list and predict

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().models_versions_create(...).doit()
let r = hub.projects().models_versions_patch(...).doit()
let r = hub.projects().operations_get(...).doit()
let r = hub.projects().models_versions_delete(...).doit()
let r = hub.projects().models_delete(...).doit()
let r = hub.projects().models_patch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-ml1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_ml1 as ml1;
use ml1::GoogleCloudMlV1__Version;
use ml1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use ml1::CloudMachineLearningEngine;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudMachineLearningEngine::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = GoogleCloudMlV1__Version::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().models_versions_patch(req, "name")
             .update_mask("sed")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Chunk

CloudMachineLearningEngine
Central instance to access all CloudMachineLearningEngine related resource activities

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GoogleApi__HttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page.

GoogleCloudMlV1__BuiltInAlgorithmOutput
Represents output related to a built-in algorithm Job.

GoogleCloudMlV1__ReplicaConfig
Represents the configuration for a replica in a cluster.

GoogleCloudMlV1__ListJobsResponse
Response message for the ListJobs method.

GoogleCloudMlV1__Config
There is no detailed description.

GoogleCloudMlV1__ListVersionsResponse
Response message for the ListVersions method.

GoogleCloudMlV1__SetDefaultVersionRequest
Request message for the SetDefaultVersion request.

GoogleCloudMlV1__Location
There is no detailed description.

GoogleCloudMlV1__HyperparameterOutput
Represents the result of a single hyperparameter tuning trial from a training job. The TrainingOutput object that is returned on successful completion of a training job with hyperparameter tuning includes a list of HyperparameterOutput objects, one for each successful trial.

GoogleCloudMlV1__HyperparameterSpec
Represents a set of hyperparameters to optimize.

GoogleCloudMlV1_HyperparameterOutput_HyperparameterMetric
An observed value of a metric.

GoogleCloudMlV1__Model
Represents a machine learning solution.

GoogleCloudMlV1__ParameterSpec
Represents a single hyperparameter to optimize.

GoogleCloudMlV1__TrainingInput
Represents input parameters for a training job. When using the gcloud command to submit your training job, you can specify the input parameters as command-line arguments and/or in a YAML configuration file referenced from the --config command-line argument. For details, see the guide to submitting a training job.

GoogleCloudMlV1__PredictRequest
Request for predictions to be issued against a trained model.

GoogleCloudMlV1__ManualScaling
Options for manually scaling a model.

GoogleCloudMlV1__ListModelsResponse
Response message for the ListModels method.

GoogleCloudMlV1__AcceleratorConfig
Represents a hardware accelerator request config.

GoogleCloudMlV1__CancelJobRequest
Request message for the CancelJob method.

GoogleCloudMlV1__PredictionOutput
Represents results of a prediction job.

GoogleCloudMlV1__TrainingOutput
Represents results of a training job. Output only.

GoogleCloudMlV1__GetConfigResponse
Returns service account information associated with a project.

GoogleCloudMlV1__Job
Represents a training or prediction job.

GoogleCloudMlV1__PredictionInput
Represents input parameters for a prediction job.

GoogleCloudMlV1__ListLocationsResponse
There is no detailed description.

GoogleCloudMlV1__Version
Represents a version of the model.

GoogleCloudMlV1__Capability
There is no detailed description.

GoogleCloudMlV1__AutoScaling
Options for automatically scaling a model.

GoogleIamV1__Policy
Defines an Identity and Access Management (IAM) policy. It is used to specify access control policies for Cloud Platform resources.

GoogleIamV1__TestIamPermissionsRequest
Request message for TestIamPermissions method.

GoogleIamV1__AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs.

GoogleIamV1__AuditLogConfig
Provides the configuration for logging a type of permissions. Example:

GoogleIamV1__TestIamPermissionsResponse
Response message for TestIamPermissions method.

GoogleIamV1__SetIamPolicyRequest
Request message for SetIamPolicy method.

GoogleIamV1__Binding
Associates members with a role.

GoogleLongrunning__ListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunning__Operation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobuf__Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

GoogleRpc__Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

GoogleType__Expr
Represents an expression text. Example:

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ProjectGetConfigCall
Get the service account information associated with your project. You need this information in order to grant the service account permissions for the Google Cloud Storage location where you put your model training code for training the model with Google Cloud Machine Learning.

ProjectJobCancelCall
Cancels a running job.

ProjectJobCreateCall
Creates a training or a batch prediction job.

ProjectJobGetCall
Describes a job.

ProjectJobGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectJobListCall
Lists the jobs in the project.

ProjectJobPatchCall
Updates a specific job resource.

ProjectJobSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy.

ProjectJobTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.

ProjectLocationGetCall
Get the complete list of CMLE capabilities in a location, along with their location-specific properties.

ProjectLocationListCall
List all locations that provides at least one type of CMLE capability.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudMachineLearningEngine hub.

ProjectModelCreateCall
Creates a model which will later contain one or more versions.

ProjectModelDeleteCall
Deletes a model.

ProjectModelGetCall
Gets information about a model, including its name, the description (if set), and the default version (if at least one version of the model has been deployed).

ProjectModelGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectModelListCall
Lists the models in a project.

ProjectModelPatchCall
Updates a specific model resource.

ProjectModelSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy.

ProjectModelTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.

ProjectModelVersionCreateCall
Creates a new version of a model from a trained TensorFlow model.

ProjectModelVersionDeleteCall
Deletes a model version.

ProjectModelVersionGetCall
Gets information about a model version.

ProjectModelVersionListCall
Gets basic information about all the versions of a model.

ProjectModelVersionPatchCall
Updates the specified Version resource.

ProjectModelVersionSetDefaultCall
Designates a version to be the default for the model.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

ProjectPredictCall
Performs prediction on the data in the request. AI Platform implements a custom predict verb on top of an HTTP POST method.
For details of the request and response format, see the guide to the predict request format.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.